Convert waitTime to float

+++ State/StatePriority.cs
@@ -1,1 +1,1 @@
-﻿public record struct StatePriority(float TimeSpent, int AvailableHP, int AvailableMoney, float heuristic);
+﻿public record struct StatePriority(float TimeSpent, float AvailableHP, int AvailableMoney, float heuristic);
